Issue #5910: Extract GenotypeGVCFs engine into publicly accessible class/function
